+++ chains/node/cli/src/lib.rs
@@ -1,0 +1,5 @@
+mod cli;
+mod command;
+pub use cli::*;
+pub use command::*;
+pub use sc_cli::{Error, Result};
