+++ chains/node/src/main.rs
@@ -2,13 +2,14 @@
 #![warn(missing_docs)]
 #![warn(unused_extern_crates)]
 
-mod chain_spec;
-#[macro_use]
-mod service;
-mod cli;
-mod command;
-mod rpc;
+// mod chain_spec;
+// #[macro_use]
+// mod service;
+// mod cli;
+// mod command;
+// mod rpc;
 
 fn main() -> sc_cli::Result<()> {
-	command::run()
+	// command::run()
+	meta_cli::run()
 }
